+++ frontend/src/lib/components/metrics/TransactionRateWarningIcon.svelte
@@ -1,0 +1,13 @@
+<script lang="ts">
+  import { IconMeter } from "@dfinity/gix-components";
+</script>
+
+<div>
+  <IconMeter />
+</div>
+
+<style lang="scss">
+  .high {
+    --meter-rotate: 70deg;
+  }
+</style>
